Added GoogleMaps SDK. Created base classes for map.

+++ app/src/main/java/ru/malcolmxio/places/domain/model/country/Coordinates.kt
@@ -1,6 +1,6 @@
 package ru.malcolmxio.places.domain.model.country
 
 data class Coordinates(
-    val lat: Long,
-    val lon: Long
+    val lat: Double,
+    val lon: Double
 )
